share HTTP_STATUS_CODE between server and client
